Move internal/otlp to model/otlp
